feat: update types, separate config
