+++ mccn/_types.py
@@ -1,5 +1,6 @@
 from __future__ import annotations
 
+from ast import Num
 from dataclasses import dataclass, field
 from typing import TYPE_CHECKING, Callable, Literal
 
@@ -8,10 +9,8 @@ from pyproj.crs.crs import CRS
 
 if TYPE_CHECKING:
     import datetime
-    from collections.abc import Mapping
 
     import pystac
-    from odc.geo.geobox import GeoBox
     from stac_generator.core import (
         PointConfig,
         RasterConfig,
@@ -31,7 +30,9 @@ _MergeMethods = (
     | Callable[[np.ndarray], float]
 )
 MergeMethods = _MergeMethods | dict[str, _MergeMethods]
-
+Number_T = int | float
+Resolution_T = Number_T | tuple[Number_T, Number_T]
+Shape_T = int | tuple[int, int]
 BBox_T = tuple[float, float, float, float]
 CRS_T = str | int | CRS
 AnchorPos_T = Literal["center", "edge", "floating", "default"] | tuple[float, float]
@@ -84,43 +85,3 @@ class ParsedRaster(ParsedItem):
     """Band aliasing - derived from eobands common name"""
     config: RasterConfig
     """STAC Generator config - raster type"""
-
-
-@dataclass
-class FilterConfig:
-    """The config that describes the extent of the cube"""
-
-    geobox: GeoBox
-    """Spatial extent"""
-    start_ts: datetime.datetime | None = None
-    """Temporal extent - start"""
-    end_ts: datetime.datetime | None = None
-    """Temporal extent - end"""
-    bands: set[str] | None = None
-    """Bands to be loaded"""
-
-
-@dataclass
-class CubeConfig:
-    """The config that describes the datacube coordinates"""
-
-    x_coord: str = "lon"
-    """Name of the x coordinate in the datacube"""
-    y_coord: str = "lat"
-    """Name of the y coordinate in the datacube"""
-    t_coord: str = "time"
-    """Name of the time coordinate in the datacube"""
-    z_coord: str = "alt"
-    """Name of the altitude coordinate in the datacube"""
-    use_z: bool = False
-    """Whether to use the altitude coordinate as an axis"""
-
-
-@dataclass
-class ProcessConfig:
-    """The config that describes data transformation and column renaming before data is loaded to the final datacube"""
-
-    rename_bands: Mapping[str, str] | None = None
-    """Mapping between original to renamed bands"""
-    process_bands: Mapping[str, Callable] | None = None
-    """Mapping between band name and transformation to be applied to the band"""
